fix bug: in rows completed, colors are wrongly replaced
The algorithm copies the previous row and set there instead of the
completed row, but the color was stored in the completed row so the
color didn't take the previous row color blocks. Now each block has a
color so when copies the previous row it has the color also for the
replaced row.